Remove convert method to OrderedDict
The generic one in Julia stdlib does the job, except for the deprecation
warning shown when converting from an unordered dict (such as `Base.Dict`)
to an `OrderedDict{K,V}` type.

This avoids potential invalidations in packages. There also is no explanation
as to why it makes sense to forbids such conversions.