Fix replace for LittleSet
Julia v1.7+ supports `replace` for tuples so had
to manually implement this here. Also removed
inference testing for the current approach since it
uses a temporary `Vector` to store results. This
is probably a sold approach for any frozen set of
tuples over 32, but it would be good to optimize
for smaller cases in the future.